Handle duplicate header names and values
If an HTTPHeaderMatch contains entries with equivalent header names, we will only configure
the first entry. We will ignore all other entries with that name. Header names are
case-insensitive, so "Foo" is equivalent to "foo".

If an HTTP request has repeated headers, nginx will merge the values into a comma delimited string.
For example, -H "foo:bar" -H "foo:baz" becomes {"foo":"bar,baz"} in the headersIn object.
In this case, a request satisfies the header match {"foo":"bar"} if one of the values specified
in the request header "foo" is "bar".